move to CIRGenOpenCL.cpp
